Backend/requirements.txt
google-generativeai
requests
pydantic
uvicorn
fastapi
python-dotenv
python-multipart